requirements.txt
langgraph
langchain
langchain-openai
psycopg2-binary
python-dotenv
PyMuPDF
reportlab
